[ScrollTimeline] Run tests on cc if threaded compositing is enabled
The scroller in the updated tests are not composited by default. When we
run the tests with the virtual/threaded test suite, the tests can not
properly run on the compositor because cc::ScrollTimeline is considered
inactive, i.e. scroller not in ScrollTree.

This patch forces compositing the scrollers if the animation needs to
play.

Bug:None
Change-Id: I0e379abb071c9bffc58f2167f4cac9f92edbb9da